Separate out the user profile form to a component
